+++ compiler/src/main/kotlin/com/r0adkll/kimchi/processors/ContributesBindingSymbolProcessor.kt
@@ -10,8 +10,11 @@ import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.r0adkll.kimchi.HINT_BINDING_PACKAGE
 import com.r0adkll.kimchi.annotations.ContributesBinding
 import com.r0adkll.kimchi.annotations.ContributesBindingAnnotation
+import com.r0adkll.kimchi.util.KimchiException
+import com.r0adkll.kimchi.util.ksp.hasAnnotation
 import com.squareup.kotlinpoet.ClassName
 import kotlin.reflect.KClass
+import me.tatarka.inject.annotations.Inject
 
 internal class ContributesBindingSymbolProcessor(
   env: SymbolProcessorEnvironment,
@@ -29,5 +32,19 @@ internal class ContributesBindingSymbolProcessor(
 
   override fun getScope(element: KSClassDeclaration): ClassName {
     return ContributesBindingAnnotation.from(element).scope
+  }
+
+  override fun validate(element: KSClassDeclaration) {
+    val hasInjectAnnotation = element.hasAnnotation(Inject::class)
+    val hasInjectedParameters = element.primaryConstructor
+      ?.parameters
+      ?.isNotEmpty() == true
+
+    if (hasInjectedParameters && !hasInjectAnnotation) {
+      throw KimchiException(
+        "@ContributesBinding annotated classes with injected constructor parameters require an '@Inject' annotation",
+        element,
+      )
+    }
   }
 }