+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Other Notes & Contributions
 -->
 
+### Fixed
+
+- Fixed issue where @ContributesBinding elements with constructor injections and no @Inject annotation
+
 ## [0.2.0] - 2024-08-24
 
 ### Added